:lipstick: 커리어 페이지 디자인 및 내용 수정
