+++ src/components/Navbar.jsx
@@ -148,7 +148,7 @@ const Navbar = ({ darkMode, setDarkMode }) => {
 
   const navItems = [
     { name: 'Home', href: '#home' },
-    { name: 'About', href: '#about' },
+    { name: 'Career', href: '#career' },
     { name: 'Skills', href: '#skills' },
     { name: 'Projects', href: '#projects' },
     { name: 'Contact', href: '#contact' },
